added delete receipt api & fully tested

+++ playground/core/services/classes/receipt_service.py
@@ -42,6 +42,11 @@ class ReceiptService:
     def delete(self, receipt_id: str, shift_service: IShiftService) -> None:
         if not self.receiptRepo.contains_receipt(receipt_id):
             raise ValueError(f"Receipt with id {receipt_id} does not exist.")
+        if self.receiptRepo.get_receipt(receipt_id).status == "closed":
+            raise ValueError("Receipt is already Closed.")
+        shift_service.remove_receipt(
+            receipt_id, self.receiptRepo.get_receipt(receipt_id).shift_id
+        )
         self.receiptRepo.delete_receipt(receipt_id)
 
     def get(self, receipt_id: str) -> Receipt:
